+++ Packages/ai.muna.muna/Runtime/Beta/OpenAI/ImageService.cs
@@ -1,0 +1,16 @@
+/* 
+*   Muna
+*   Copyright © 2026 NatML Inc. All rights reserved.
+*/
+
+#nullable enable
+
+namespace Muna.Beta.OpenAI {
+
+    /// <summary>
+    /// Create images.
+    /// </summary>
+    public sealed class ImageService {
+
+    }
+}