+++ Packages/ai.muna.muna/Runtime/Beta/Types/ValueUtils.cs
@@ -14,29 +14,6 @@ namespace Muna.Beta {
     using System.Runtime.Serialization;
     using System.Text;
 
-    /// <summary>
-    /// Remote prediction value.
-    /// </summary>
-    [Preserve, Serializable]
-    public sealed class RemoteValue {
-
-        /// <summary>
-        /// Value URL.
-        /// </summary>
-        public string? data;
-
-        /// <summary>
-        /// Value type.
-        /// </summary>
-        public Dtype dtype;
-
-        /// <summary>
-        /// Value shape.
-        /// This is `null` if shape information is not available or applicable.
-        /// </summary>
-        public int[]? shape;
-    }
-
     internal static class ValueUtils {
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]